pyproject.toml
[project]
name = "langchain-mongodb-monorepo"
version = "0.1.0"
description = "LangChain MongoDB mono-repo"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = []

[dependency-groups]
dev = [
    "autodoc-pydantic>=2.2.0",
    "pydata-sphinx-theme>=0.16.1",
    "sphinx-design>=0.6.1",
    "sphinx>=7.4.7",
    "sphinx-copybutton>=0.5.2",
    "toml>=0.10.2",
    "langchain-core>=0.3.30",
    "sphinxcontrib-googleanalytics>=0.4",
    "langchain-mongodb",
    "langgraph-checkpoint-mongodb",
    "langgraph-store-mongodb",
    "langchain-community>=0.3.14",
    "myst-parser>=3.0.1",
    "ipython>=8.18.1",
]

[tool.uv.sources]
langchain-mongodb = { path = "libs/langchain-mongodb", editable = true }
langgraph-checkpoint-mongodb = { path = "libs/langgraph-checkpoint-mongodb", editable = true }
langgraph-store-mongodb = { path = "libs/langgraph-store-mongodb", editable = true }
